DUGK-67 Removed 'Workflows.tsx' and renamed '...Data.ts' (#45)
